// frontend/src/components/__generated__/getRolePermissions.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: getRolePermissions
// ====================================================

export interface getRolePermissions_getRolePermissions {
  __typename: "RolePermission";
  name: string;
}

export interface getRolePermissions {
  getRolePermissions: getRolePermissions_getRolePermissions[] | null;
}